requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
pydantic-settings
alembic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
multipart
email-validator